+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and follow [semantic versioning](https://semver.org/) for our releases.
 
 ### Fixed
 
+- [#243](https://github.com/EspressoSystems/jellyfish/pull/243) fixes bug in MerkleTreeGadget implementation for SparseRescueMerkleTree.
+
 ### Added
 
 - [#233](https://github.com/EspressoSystems/jellyfish/pull/233) BLS aggregation APIs